backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
sqlalchemy>=2.0.0
python-multipart>=0.0.6
psycopg2-binary>=2.9.0; sys_platform != 'win32'
pyodbc>=5.0.0; sys_platform == 'win32'
google-cloud-storage>=2.10.0; extra == "gcs"

